backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
clerk-backend-api==3.3.1
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.0.0